requirements.txt
streamlit
langchain
langchain-community
python-dotenv
pypdf2
python-docx
sentence-transformers
faiss-cpu
google-generativeai
fpdf
